+++ zakozhurnikova.kristina/P4/file_operation.cpp
@@ -1,8 +1,9 @@
 #include "file_operation.h"
 #include <ostream>
 #include <exception>
+#include <istream>
 
-int zakozhurnikova::checkArguments(int argc, const char **const argv)
+int zakozhurnikova::checkArguments(int argc, const char *const *const argv)
 {
   if (argc != 4)
   {
@@ -37,4 +38,18 @@ void zakozhurnikova::writeToDest(std::ostream& out, const int *matrix, size_t ro
     out << matrix[i] << ' ';
   }
   out << matrix[rows * cols - 1];
+}
+void zakozhurnikova::writeToMatrix(std::istream& in, int *matrix, size_t rows, size_t cols)
+{
+  for (size_t i = 0; i < rows; i++)
+  {
+    for (size_t j = 0; j < cols; j++)
+    {
+      in >> matrix[i * cols + j];
+      if (!in)
+      {
+        throw std::runtime_error("Read from file error");
+      }
+    }
+  }
 }
