+++ zakozhurnikova.kristina/P4/file_operation.h
@@ -5,7 +5,8 @@
 
 namespace zakozhurnikova
 {
-  int checkArguments(int argc, const char **const argv);
+  int checkArguments(int argc, const char *const *const argv);
   void writeToDest(std::ostream& out, const int *matrix, size_t rows, size_t cols);
+  void writeToMatrix(std::istream& in, int *matrix, size_t rows, size_t cols);
 }
 #endif
