+++ zakozhurnikova.kristina/P4/matrix_operation.h
@@ -7,6 +7,5 @@ namespace zakozhurnikova
 {
   void fillMatrix(int *matrix, size_t rows, size_t cols);
   void substractMatrix(int *original, const int *substract, size_t rows, size_t cols);
-  void writeToMatrix(std::istream& in, int *matrix, size_t rows, size_t cols);
 }
 #endif
